css/css-position/position-sticky-bottom.html

<!DOCTYPE html>
<title>position:sticky elements should respect the bottom constraint</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that position:sticky elements obey their bottom anchor after scrolling" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.scroller {
  position: relative;
  width: 100px;
  height: 200px;
  overflow-x: hidden;
  overflow-y: scroll;
}

.contents {
  height: 500px;
}

.prepadding {
  height: 200px;
}

.container {
  height: 200px;
}

.filler {
  height: 100px;
}

.sticky {
  background-color: green;
  position: sticky;
  bottom: 25px;
}

.box {
  width: 100%;
  height: 100px;
}
</style>

<div id="scroller1" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="filler"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<div id="scroller2" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="filler"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<div id="scroller3" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="filler"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<script>
test(() => {
  const scroller = document.getElementById('scroller1');
  const prepadding = scroller.querySelector('.prepadding');
  const filler = scroller.querySelector('.filler');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollTop = 275;

  const nonStickyTopY = prepadding.clientHeight + filler.clientHeight;
  assert_equals(sticky.offsetTop, nonStickyTopY);
}, 'before reaching the sticking point the sticky box should not be offset');

test(() => {
  const scroller = document.getElementById('scroller2');
  const prepadding = scroller.querySelector('.prepadding');
  const filler = scroller.querySelector('.filler');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollTop = 175;

  const nonStickyTopY = prepadding.clientHeight + filler.clientHeight;
  const nonStickyBottomY = nonStickyTopY + sticky.clientHeight;
  const targetBottomY = scroller.clientHeight + scroller.scrollTop - 25;
  const stickyOffset = nonStickyBottomY - targetBottomY;

  assert_equals(sticky.offsetTop, nonStickyTopY - stickyOffset);
}, 'after reaching the sticking point the sticky box should be offset');

test(() => {
  const scroller = document.getElementById('scroller3');
  const container = scroller.querySelector('.container');
  const prepadding = scroller.querySelector('.prepadding');
  const filler = scroller.querySelector('.filler');
  const sticky = container.querySelector('.sticky');

  scroller.scrollTop = 100;

  // The sticky is at the bottom, so hax.
  const nonStickyTopY = prepadding.clientHeight + filler.clientHeight;
  const maxOffsetInContainer = container.clientHeight - sticky.clientHeight;

  assert_equals(sticky.offsetTop, nonStickyTopY - maxOffsetInContainer);
}, 'the sticky box should not be pushed outside its containing block');
</script>
